+++ CHANGELOG.md
@@ -3,3 +3,5 @@
 ## 0.1.x
 
 ### 0.1.0
+
+- middleware: Sentry by @poteto0 in #1